Add files via upload
This is update .8. This update has been a long time coming, and I regret not putting it out sooner.
I was busy with a lot of personal stuff so this went onto the sidelines for a while. Anyway some changes
include a new explore function, proper default settings and support for installing and updating the application 
via make. Now you can just make install to install the application. No more tinkering with file. ENJOY!!