issue-fix-update
This update is in response to the numerous reported.
Hopefully this fixes all issues currently reported.
One thing to note is that this update also disables all
functions having two arguments, since the explorer 
replaces it nicely. So, until further notice these functions
will remain disabled.

Patch notes:
1. All functions in two.cpp has been disabled
and will remain as such until further notice.

2.Help file is updated with new changes.

3. Provided the "del" command to explorer
to delete files in project.

4.Provided the "back" command to explorer
in order to go back to the previous directory
in the project.

5.Issue regarding saving project paths have 
been fixed.